+++ src/main/java/gravestone/structures/catacombs/components/Bridge.java
@@ -77,7 +77,7 @@ public class Bridge extends CatacombsBaseComponent {
         this.fillWithRandomizedBlocks(world, boundingBox, 10, 9, 7, 11, 10, 7, false, random, getCemeteryCatacombsStones());
         this.fillWithRandomizedBlocks(world, boundingBox, 10, 9, 7, 11, 10, 7, false, random, getCemeteryCatacombsStones());
         // graves
-        byte graveType = GraveStoneHelper.getGraveType(world, new BlockPos(this.getXWithOffset(0, 0), this.getYWithOffset(0), this.getZWithOffset(0, 0)), random, EnumGraveType.PLAYER_GRAVES);
+        int graveType = GraveStoneHelper.getGraveType(world, new BlockPos(this.getXWithOffset(0, 0), this.getYWithOffset(0), this.getZWithOffset(0, 0)), random, EnumGraveType.PLAYER_GRAVES);
         Item sword = GraveStoneHelper.getRandomSwordForGeneration(graveType, random);
 
         IBlockState graveState = GSBlock.graveStone.getDefaultState();
